+++ NEWS.md
@@ -1,8 +1,9 @@
 # Change Log
 
-## Unreleased
+## [2.1.0] - 2019-04-22
 ### Changed
 - alpha of 0 line in difference smooth plot (now set to 0.5)
+- prepared for CRAN first release
 
 ### Fixed
 - missing import in `plot_difference()`